U540: sbi_init to S-Mode and add mem region for EDK2
Port a5ac63096ca5da795042baf650170643fe219cab to U540

Description:

To remove our custom patches from OpenSBI we try to make its checks
happy. It's no longer allows to stay in M-Mode during sbi_init, so we'll
switch to S-Mode directly.

However, when doing that it checks that the next address is executable.
Because OpenSBI marks the FW region without any attributes, it's not.
Therefore we have to add our own region with RWX set.
This makes the check in OpenSBI happy and should also set the PMP setting.
However it doesn't because we have to add another region to make OpenSBI
happy. The FW region. This must be exactly the same as the FW region
hardcoded in OpenSBI and it will override the PMP settings in our EDK2
region.